// src/generated/api/models/NotificationDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserSummaryDto } from './UserSummaryDto';
import {
    UserSummaryDtoFromJSON,
    UserSummaryDtoFromJSONTyped,
    UserSummaryDtoToJSON,
    UserSummaryDtoToJSONTyped,
} from './UserSummaryDto';

/**
 * 
 * @export
 * @interface NotificationDto
 */
export interface NotificationDto {
    /**
     * 
     * @type {number}
     * @memberof NotificationDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof NotificationDto
     */
    type?: NotificationDtoTypeEnum;
    /**
     * 
     * @type {UserSummaryDto}
     * @memberof NotificationDto
     */
    actor?: UserSummaryDto;
    /**
     * 
     * @type {number}
     * @memberof NotificationDto
     */
    postId?: number;
    /**
     * 
     * @type {number}
     * @memberof NotificationDto
     */
    commentId?: number;
    /**
     * 
     * @type {boolean}
     * @memberof NotificationDto
     */
    read?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof NotificationDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {string}
     * @memberof NotificationDto
     */
    message?: string;
}


/**
 * @export
 */
export const NotificationDtoTypeEnum = {
    NewPost: 'NEW_POST',
    NewLike: 'NEW_LIKE',
    NewComment: 'NEW_COMMENT',
    NewFollow: 'NEW_FOLLOW'
} as const;
export type NotificationDtoTypeEnum = typeof NotificationDtoTypeEnum[keyof typeof NotificationDtoTypeEnum];


/**
 * Check if a given object implements the NotificationDto interface.
 */
export function instanceOfNotificationDto(value: object): value is NotificationDto {
    return true;
}

export function NotificationDtoFromJSON(json: any): NotificationDto {
    return NotificationDtoFromJSONTyped(json, false);
}

export function NotificationDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NotificationDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'type': json['type'] == null ? undefined : json['type'],
        'actor': json['actor'] == null ? undefined : UserSummaryDtoFromJSON(json['actor']),
        'postId': json['postId'] == null ? undefined : json['postId'],
        'commentId': json['commentId'] == null ? undefined : json['commentId'],
        'read': json['read'] == null ? undefined : json['read'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'message': json['message'] == null ? undefined : json['message'],
    };
}

export function NotificationDtoToJSON(json: any): NotificationDto {
    return NotificationDtoToJSONTyped(json, false);
}

export function NotificationDtoToJSONTyped(value?: NotificationDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'type': value['type'],
        'actor': UserSummaryDtoToJSON(value['actor']),
        'postId': value['postId'],
        'commentId': value['commentId'],
        'read': value['read'],
        'createdAt': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
        'message': value['message'],
    };
}